updated extractor and background

+++ src/main/java/com/eso_encore/launcher/updater/InitialInstallTask.xtend
@@ -1,16 +1,17 @@
 package com.eso_encore.launcher.updater
 
+import com.eso_encore.launcher.Action
+import com.eso_encore.launcher.Launcher
 import com.eso_encore.launcher.service.WebsiteService
-import java.io.FileOutputStream
-import java.io.RandomAccessFile
 import java.net.HttpURLConnection
 import java.nio.file.Files
 import java.nio.file.Paths
+import java.time.Duration
+import java.time.ZonedDateTime
+import java.time.temporal.ChronoUnit
 import javafx.application.Platform
 import javafx.concurrent.Task
-import net.sf.sevenzipjbinding.ArchiveFormat
-import net.sf.sevenzipjbinding.SevenZip
-import net.sf.sevenzipjbinding.impl.RandomAccessFileInStream
+import net.sf.sevenzipjbinding.ExtractOperationResult
 import org.apache.commons.io.FileUtils
 import org.apache.logging.log4j.LogManager
 import org.eclipse.xtend.lib.annotations.FinalFieldsConstructor
@@ -26,62 +27,56 @@ class InitialInstallTask extends Task<Void> {
 	val Updater updater
 
 	override protected call() throws Exception {
+		val start = ZonedDateTime.now()
 		try {
 			updateTitle("Setting up")
 			val saveFile = Paths.get(properties.saveFile)
 			val installationDirectory = Paths.get(properties.installationDirectory)
 
-			updateMessage("Getting download size")
-			val size = website.size
-			println("size " + size)
-			val sizeString = FileUtils.byteCountToDisplaySize(size)
-			updateMessage("Getting version")
-			val version = website.version
+			if (properties.currentAction != Action.INITIAL_EXTRACT) {
+				Launcher.save(properties.withCurrentAction(Action.INITIAL_DOWNLOAD))
+				updateMessage("Getting download size")
+				val size = website.size
+				println("size " + size)
+				val sizeString = FileUtils.byteCountToDisplaySize(size)
+				updateMessage("Getting version")
+				val version = website.version
 
-			updateTitle('''Downloading update «version» to «saveFile»''')
-			updateMessage("")
-			val url = website.getUrl("/api/download")
-			val conn = url.openConnection as HttpURLConnection
-			conn.requestMethod = "GET"
+				updateTitle('''Downloading update «version» to «saveFile»''')
+				updateMessage("")
+				val url = website.getUrl("/api/download")
+				val conn = url.openConnection as HttpURLConnection
+				conn.requestMethod = "GET"
 
-			if(Files.exists(saveFile)) {
-				Files.delete(saveFile)
+				if (Files.exists(saveFile)) {
+					Files.delete(saveFile)
+				}
+				updater.downloadWithProgress(conn, saveFile.toFile) [ bytesRead |
+					updateProgress(bytesRead, size)
+					updateMessage(
+						FileUtils.byteCountToDisplaySize(bytesRead) + "/" + sizeString + "\t(" + bytesRead + "b/" +
+							size + "b)\t" + (bytesRead.doubleValue / size * 100).floatValue + "%")
+				]
+				println("Downloaded")
 			}
-			updater.downloadWithProgress(conn, saveFile.toFile) [ bytesRead |
-				updateProgress(bytesRead, size)
-				updateMessage(
-					FileUtils.byteCountToDisplaySize(bytesRead) + "/" + sizeString + "\t(" + bytesRead + "b/" + size +
-						"b)\t" + (bytesRead.doubleValue / size * 100).floatValue + "%")
-			]
-			println("Downloaded")
 
 			updateTitle("Extracting new installation")
 			updateMessage("")
 			updateProgress(-1, 1)
+			Launcher.save(properties.withCurrentAction(Action.INITIAL_EXTRACT))
 			println("Extracting")
 			installationDirectory.toFile.mkdirs()
-			
-			val randomAccessFile = new RandomAccessFile(saveFile.toFile, "r")
-			val archive = SevenZip.openInArchive(ArchiveFormat.SEVEN_ZIP, new RandomAccessFileInStream(randomAccessFile))
-			val simpleArchive = archive.simpleInterface
-			val itemCount = simpleArchive.numberOfItems
-			simpleArchive.archiveItems.forEach[it,index|
-				val outPath = Paths.get(properties.installationDirectory, it.path)
-				updateMessage("Extracting " + it.path)
-				if(it.isFolder) {
-					outPath.toFile().mkdirs()
-				} else {
-					try(val out = new FileOutputStream(outPath.toFile)) {
-						it.extractSlow[data|
-							out.write(data)
-							return data.length
-						]
-					}
+
+			SevenZipExtractor.extract(saveFile, Paths.get(properties.installationDirectory)) [index, total, path, result|
+				if(result == ExtractOperationResult.OK) {
+					updateMessage("Extracted "+path+"\t"+index+"/"+total)
+					updateProgress(index, total)
 				}
-				updateProgress(index+1, itemCount)
 			]
-			
-			updateTitle("Finished installing")
+
+			Launcher.save(properties.withCurrentAction(Action.NONE))
+			val duration = ChronoUnit.SECONDS.between(start, ZonedDateTime.now)
+			updateTitle("Finished installing in " + Duration.ofSeconds(duration).humanReadableFormat)
 			updateMessage("")
 		} catch (Exception e) {
 			log.error("Failed to install initial", e)
@@ -90,6 +85,10 @@ class InitialInstallTask extends Task<Void> {
 		}
 
 		return null
+	}
+
+	def static String humanReadableFormat(Duration duration) {
+		return duration.toString().substring(2).replaceAll("(\\d[HMS])(?!$)", "$1 ").toLowerCase()
 	}
 
 	override protected updateTitle(String title) {
